Exotic Option Pricing-Asian and Lookback
Asian Option--

Asian Option payoff depends on the Average value of the underlying asset over some period before expiry. In our current problem we have taken the Average over the entire period before expiry as an Arithmetic Average based on daily closing price.
The Average value is compared using a Fixed Strike to value Asian Option for both calls and puts with Fixed Strike and with Floating Strike to value Asian option with Floating Strike to derive the option price and compared under different scenarios by varying the data.

Lookback Option--

Lookback Option payoff depends on the realized maximum or minimum of the underlying asset adjusted with the strike price over some period before expiry. In our current problem we have calculated the Maximum and Minimum value over the entire period before expiry and compared using a fixed strike and a floating strike to determine the value for both calls and puts.